pyproject.toml
[project]
name = "carsensor-backend"
version = "0.1.0"
requires-python = ">=3.14"
dependencies = [
    "beautifulsoup4[lxml]>=4.14.3",
    "boto3>=1.42.72",
    "celery>=5.6.2",
    "dj-database-url>=3.1.2",
    "django>=6.0.3",
    "django-celery-beat>=2.9.0",
    "django-celery-results>=2.6.0",
    "django-cors-headers>=4.9.0",
    "django-environ>=0.13.0",
    "django-filter>=25.2",
    "django-redis>=6.0.0",
    "django-storages[s3]>=1.14.6",
    "djangorestframework>=3.17.0",
    "djangorestframework-simplejwt>=5.5.1",
    "drf-spectacular>=0.29.0",
    "gunicorn>=25.1.0",
    "httpx>=0.28.1",
    "pillow>=12.1.1",
    "psycopg[binary,pool]>=3.3.3",
    "uvicorn>=0.42.0",
    "whitenoise>=6.12.0",
]

[dependency-groups]
dev = [
    "django-debug-toolbar>=6.2.0",
    "django-extensions>=4.1",
    "django-silk>=5.5.0",
    "django-stubs>=6.0.1",
    "factory-boy>=3.3.3",
    "faker>=40.11.0",
    "ipython>=9.11.0",
    "mypy>=1.19.1",
    "pytest>=9.0.2",
    "pytest-cov>=7.0.0",
    "pytest-django>=4.12.0",
]

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "config.settings.testing"
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v --tb=short"
